[D-B-F9]platform反馈（问卷）功能实现 1110

+++ src/main/java/xyz/eulix/platform/services/mgtboard/dto/QuestionnaireReq.java
@@ -28,5 +28,5 @@ public class QuestionnaireReq {
 
     @NotNull
     @Schema(required = true, description = "第三方问卷id")
-    private Integer payloadSurveyId;
+    private Long payloadSurveyId;
 }
